python/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
diffusers==0.29.2
transformers==4.41.2
accelerate==0.31.0
huggingface_hub==0.23.4
sam2==1.1.0
trimesh==4.4.0
numpy==1.26.4
Pillow==10.3.0
scipy==1.13.1